+++ data/statistics_cricket.go
@@ -2,6 +2,7 @@ package data
 
 import (
 	"database/sql"
+	"fmt"
 
 	"github.com/guregu/null"
 
@@ -287,4 +288,21 @@ func CalculateCricketStatistics(legID int) (map[int]*models.StatisticsCricket, e
 		stat.Rounds = round
 	}
 	return statisticsMap, nil
+}
+
+// RecalculateCricketStatistics will recaulcate statistics for Cricket matches
+func RecalculateCricketStatistics(legs []int) ([]string, error) {
+	queries := make([]string, 0)
+	for _, legID := range legs {
+		stats, err := CalculateCricketStatistics(legID)
+		if err != nil {
+			return nil, err
+		}
+		for playerID, stat := range stats {
+
+			queries = append(queries, fmt.Sprintf(`UPDATE statistics_cricket SET total_marks = %d, rounds = %d, score = %d, first_nine_marks = %d mpr = %f, first_nine_mpr = %f, marks5 = %d, marks6 = %d, marks7 = %d, marks8 = %d, marks9 = %d, WHERE leg_id = %d AND player_id = %d;`,
+				stat.TotalMarks, stat.Rounds, stat.Score.Int64, stat.FirstNineMarks, stat.MPR, stat.FirstNineMPR, stat.Marks5, stat.Marks6, stat.Marks7, stat.Marks8, stat.Marks9, legID, playerID))
+		}
+	}
+	return queries, nil
 }