+++ data/v2/player.go
@@ -50,35 +50,35 @@ func GetMatchesPlayedPerPlayer() (map[int]*models.Player, error) {
 	rows, err := models.DB.Query(`
 		SELECT
 			player_id,
-			MAX(matches_played) AS 'matches_played',
-			MAX(matches_won) AS 'matches_won',
-			MAX(legs_played) AS 'legs_played',
-			MAX(legs_won) AS 'legs_won'
+			MAX(matches_played) AS matches_played,
+			MAX(matches_won) AS matches_won,
+			MAX(legs_played) AS legs_played,
+			MAX(legs_won) AS legs_won
 		FROM (
 			SELECT
 				p2l.player_id,
-				COUNT(DISTINCT p2l.match_id) AS 'matches_played',
-				0 AS 'matches_won',
-				COUNT(m.id)  AS 'legs_played',
-				SUM(CASE WHEN p2l.player_id = m.winner_id THEN 1 ELSE 0 END) AS 'legs_won'
+				COUNT(DISTINCT p2l.match_id) AS matches_played,
+				SUM(CASE WHEN p2l.player_id = m.winner_id THEN 1 ELSE 0 END) AS matches_won,
+				COUNT(p2l.leg_id) AS legs_played,
+				SUM(CASE WHEN p2l.player_id = m.winner_id THEN 1 ELSE 0 END) AS legs_won
 			FROM player2leg p2l
-				JOIN leg l ON l.id = p2l.leg_id
 				JOIN matches m ON m.id = p2l.match_id
+				JOIN leg l ON l.id = p2l.leg_id AND l.match_id = m.id
 			WHERE l.is_finished = 1 AND m.is_abandoned = 0 AND m.is_walkover = 0
 			GROUP BY p2l.player_id
 			UNION ALL
 			SELECT
-				p2l.player_id,
-				0 AS 'matches_played',
-				COUNT(DISTINCT m.id) AS 'matches_won',
-				0 AS 'legs_played',
-				0 AS 'legs_won'
+				m.winner_id AS player_id,
+				0 AS matches_played,
+				COUNT(DISTINCT m.id) AS matches_won,
+				0 AS legs_played,
+				0 AS legs_won
 			FROM matches m
 				JOIN leg l ON l.match_id = m.id
-				JOIN player2leg p2l ON p2l.player_id = m.winner_id AND p2l.match_id = m.id
 			WHERE l.is_finished = 1 AND m.is_abandoned = 0 AND m.is_walkover = 0
 			GROUP BY m.winner_id
-		) matches
+		) AS subquery
+		WHERE player_id IS NOT NULL
 		GROUP BY player_id`)
 	if err != nil {
 		return nil, err